// src/generated/models/Mserp_inventoperationalsitev2entitiesModel.ts
export const Mserp_inventoperationalsitev2entitiesmserp_isprimaryaddressassigned = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_inventoperationalsitev2entitiesmserp_isprimaryaddressassigned = keyof typeof Mserp_inventoperationalsitev2entitiesmserp_isprimaryaddressassigned;
export const Mserp_inventoperationalsitev2entitiesmserp_isreceivingwarehouseoverrideallowed = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_inventoperationalsitev2entitiesmserp_isreceivingwarehouseoverrideallowed = keyof typeof Mserp_inventoperationalsitev2entitiesmserp_isreceivingwarehouseoverrideallowed;
export const Mserp_inventoperationalsitev2entitiesmserp_primaryaddresstimezone = {
  200000000: '_GMT_12_00_InternationalDateLineWest',
  200000001: '_GMT_11_00_CoordinatedUniversalTime_11',
  200000002: '_GMT_10_00_Hawaii',
  200000003: '_GMT_09_00_Alaska',
  200000004: '_GMT_08_00_PacificTime_US_Canada_',
  200000005: '_GMT_08_00_Tijuana_BajaCalifornia',
  200000006: '_GMT_07_00_Arizona',
  200000007: '_GMT_07_00_MountainTime_US_Canada_',
  200000008: '_GMT_07_00_Chihuahua_LaPaz_Mazatlan',
  200000009: '_GMT_06_00_CentralAmerica',
  200000010: '_GMT_06_00_CentralTime_US_Canada_',
  200000011: '_GMT_06_00_Guadalajara_MexicoCity_Monterrey',
  200000012: '_GMT_06_00_Saskatchewan',
  200000013: '_GMT_05_00_Bogota_Lima_Quito_RioBranco',
  200000014: '_GMT_05_00_Chetumal',
  200000015: '_GMT_05_00_EasternTime_US_Canada_',
  200000016: '_GMT_05_00_Indiana_East_',
  200000017: '_GMT_04_30_Caracas',
  200000018: '_GMT_04_00_Asuncion',
  200000019: '_GMT_04_00_AtlanticTime_Canada_',
  200000020: '_GMT_04_00_LaPaz',
  200000021: '_GMT_04_00_Manaus',
  200000022: '_GMT_04_00_Santiago',
  200000023: '_GMT_03_30_Newfoundland',
  200000024: '_GMT_03_00_Brasilia',
  200000025: '_GMT_03_00_BuenosAires',
  200000026: '_GMT_03_00_BuenosAires_Georgetown',
  200000027: '_GMT_03_00_Greenland',
  200000028: '_GMT_03_00_Montevideo',
  200000029: '_GMT_03_00_Salvador',
  200000030: '_GMT_02_00_Mid_Atlantic',
  200000031: '_GMT_01_00_Azores',
  200000032: '_GMT_01_00_CaboVerdeIs_',
  200000033: '_GMT_Casablanca',
  200000034: '_GMT_CoordinatedUniversalTime',
  200000035: '_GMT_Casablanca_Monrovia_Reykjavik',
  200000036: '_GMT_GreenwichMeanTime_Dublin_Edinburgh_Lisbon_London',
  200000037: '_GMT_01_00_Amsterdam_Berlin_Bern_Rome_Stockholm_Vienna',
  200000038: '_GMT_01_00_Belgrade_Bratislava_Budapest_Ljubljana_Prague',
  200000039: '_GMT_01_00_Brussels_Copenhagen_Madrid_Paris',
  200000040: '_GMT_01_00_Sarajevo_Skopje_Warsaw_Zagreb',
  200000041: '_GMT_01_00_Tripoli',
  200000042: '_GMT_01_00_WestCentralAfrica',
  200000043: '_GMT_02_00_Amman',
  200000044: '_GMT_02_00_Athens_Bucharest_Istanbul',
  200000045: '_GMT_02_00_Beirut',
  200000046: '_GMT_02_00_Minsk',
  200000047: '_GMT_02_00_Cairo',
  200000048: '_GMT_02_00_Damascus',
  200000049: '_GMT_02_00_Harare_Pretoria',
  200000050: '_GMT_02_00_Helsinki_Kyiv_Riga_Sofia_Tallinn_Vilnius',
  200000051: '_GMT_02_00_Jerusalem',
  200000052: '_GMT_02_00_Windhoek',
  200000053: '_GMT_03_00_Istanbul',
  200000054: '_GMT_02_00_Kaliningrad_RTZ1_',
  200000055: '_GMT_03_00_Baghdad',
  200000056: '_GMT_03_00_Kuwait_Riyadh',
  200000057: '_GMT_03_00_Moscow_St_Petersburg_Volgograd_RTZ2_',
  200000058: '_GMT_04_00_PortLouis',
  200000059: '_GMT_03_00_Nairobi',
  200000060: '_GMT_03_00_Tbilisi',
  200000061: '_GMT_03_30_Tehran',
  200000062: '_GMT_04_00_AbuDhabi_Muscat',
  200000063: '_GMT_04_00_Baku',
  200000064: '_GMT_04_00_Izhevsk_Samara_RTZ3_',
  200000065: '_GMT_04_00_Yerevan',
  200000066: '_GMT_04_00_Yerevan_Donotuse_',
  200000067: '_GMT_04_30_Kabul',
  200000068: '_GMT_05_00_Ekaterinburg_RTZ4_',
  200000069: '_GMT_05_00_Islamabad_Karachi',
  200000070: '_GMT_05_00_Toshkent',
  200000071: '_GMT_05_30_Chennai_Kolkata_Mumbai_NewDelhi',
  200000072: '_GMT_05_30_SriJayawardenepura',
  200000073: '_GMT_05_45_Kathmandu',
  200000074: '_GMT_06_00_Nur_Sultan_Dhaka',
  200000075: '_GMT_06_00_Dhaka',
  200000076: '_GMT_06_00_Novosibirsk_RTZ5_',
  200000077: '_GMT_06_30_Yangon_Rangoon_',
  200000078: '_GMT_07_00_Bangkok_Hanoi_Jakarta',
  200000079: '_GMT_07_00_Krasnoyarsk_RTZ6_',
  200000080: '_GMT_08_00_Beijing_Chongqing_HongKong_Urumqi',
  200000081: '_GMT_08_00_Irkutsk_RTZ7_',
  200000082: '_GMT_08_00_Ulaanbaatar',
  200000083: '_GMT_08_00_KualaLumpur_Singapore',
  200000084: '_GMT_08_00_Perth',
  200000085: '_GMT_08_00_Taipei',
  200000086: '_GMT_09_00_Osaka_Sapporo_Tokyo',
  200000087: '_GMT_09_00_Seoul',
  200000088: '_GMT_09_00_Yakutsk_RTZ8_',
  200000089: '_GMT_09_30_Adelaide',
  200000090: '_GMT_09_30_Darwin',
  200000091: '_GMT_10_00_Brisbane',
  200000092: '_GMT_10_00_Canberra_Melbourne_Sydney',
  200000093: '_GMT_10_00_Guam_PortMoresby',
  200000094: '_GMT_10_00_Hobart',
  200000095: '_GMT_10_00_Vladivostok_Magadan_RTZ9_',
  200000096: '_GMT_11_00_Chokurdakh_RTZ10_',
  200000097: '_GMT_11_00_Magadan_SolomonIs__NewCaledonia',
  200000098: '_GMT_12_00_Magadan_Donotuse_',
  200000099: '_GMT_12_00_Anadyr_Petropavlovsk_Kamchatsky_RTZ11_',
  200000100: '_GMT_12_00_Auckland_Wellington',
  200000101: '_GMT_12_00_CoordinatedUniversalTime_12',
  200000102: '_GMT_12_00_Fiji_Kamchatka_MarshallIs_',
  200000103: '_GMT_13_00_Samoa',
  200000104: '_GMT_13_00_Nuku_alofa'
} as const;
export type Mserp_inventoperationalsitev2entitiesmserp_primaryaddresstimezone = keyof typeof Mserp_inventoperationalsitev2entitiesmserp_primaryaddresstimezone;
// ...
export type Mserp_inventoperationalsitev2entitiesmserp_sitetimezone = keyof typeof Mserp_inventoperationalsitev2entitiesmserp_sitetimezone;
export const Mserp_inventoperationalsitev2entitiesmserp_willmasterplannedintrasitemovementsusetransferjournals = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_inventoperationalsitev2entitiesmserp_willmasterplannedintrasitemovementsusetransferjournals = keyof typeof Mserp_inventoperationalsitev2entitiesmserp_willmasterplannedintrasitemovementsusetransferjournals;

export interface Mserp_inventoperationalsitev2entitiesBase {
  mserp_dataareaid?: string;
  "mserp_dataAreaId_id@odata.bind"?: string;
  mserp_defaultinventorystatusid?: string;
  mserp_establishment?: number;
  mserp_establishment_bigint?: number;
  mserp_fiscalestablishmentid?: string;
  mserp_inventoperationalsitev2entityid: string;
  mserp_inventsitedefaultdimensiondisplayvalue?: string;
  mserp_isprimaryaddressassigned?: Mserp_inventoperationalsitev2entitiesmserp_isprimaryaddressassigned;
  mserp_isreceivingwarehouseoverrideallowed?: Mserp_inventoperationalsitev2entitiesmserp_isreceivingwarehouseoverrideallowed;
  mserp_orderentrydeadlinegroupid?: string;
  mserp_primaryaddressbuildingcompliment?: string;
  mserp_primaryaddresscity?: string;
  mserp_primaryaddresscityinkana?: string;
  mserp_primaryaddresscountryregionid?: string;
  mserp_primaryaddresscountyid?: string;
  mserp_primaryaddressdescription?: string;
  mserp_primaryaddressdistrictname?: string;
  mserp_primaryaddresslatitude?: number;
  mserp_primaryaddresslocationroles?: string;
  mserp_primaryaddresslocationsalestaxgroupcode?: string;
  mserp_primaryaddresslongitude?: number;
  mserp_primaryaddresspostbox?: string;
  mserp_primaryaddressstateid?: string;
  mserp_primaryaddressstreet?: string;
  mserp_primaryaddressstreetinkana?: string;
  mserp_primaryaddressstreetnumber?: string;
  mserp_primaryaddresstimezone?: Mserp_inventoperationalsitev2entitiesmserp_primaryaddresstimezone;
  mserp_primaryaddresszipcode?: string;
  mserp_siteid: string;
  mserp_sitename?: string;
  mserp_sitetimezone?: Mserp_inventoperationalsitev2entitiesmserp_sitetimezone;
  mserp_taxbranchcode?: string;
  mserp_willmasterplannedintrasitemovementsusetransferjournals?: Mserp_inventoperationalsitev2entitiesmserp_willmasterplannedintrasitemovementsusetransferjournals;
}

export interface Mserp_inventoperationalsitev2entities extends Mserp_inventoperationalsitev2entitiesBase {
  mserp_dataareaid_idname?: string;
  mserp_formattedprimaryaddress?: string;
  mserp_isprimaryaddressassignedname?: string;
  mserp_isreceivingwarehouseoverrideallowedname?: string;
  mserp_primaryaddresstimezonename?: string;
  mserp_sitetimezonename?: string;
  mserp_willmasterplannedintrasitemovementsusetransferjournalsname?: string;
  mserp_dataareaid_id?: object;
  _mserp_dataareaid_id_value?: string;
}