// src/generated/models/Mserp_inventwarehouseentitiesModel.ts
export interface Mserp_inventwarehouseentitiesBase {
// ...
  mserp_warehouseid: string;
// ...
}

export interface Mserp_inventwarehouseentities extends Mserp_inventwarehouseentitiesBase {
  mserp_areadvancedwarehousemanagementprocessesenabledname?: string;
  mserp_areitemscoverageplannedmanuallyname?: string;
  mserp_arelaborstandardsallowedname?: string;
  mserp_arepickinglistsdeliverymodespecificname?: string;
  mserp_arepickinglistsshipmentspecificonlyname?: string;
  mserp_arewarehouselocationcheckdigitsuniquename?: string;
  mserp_autoupdateshipmentrulename?: string;
  mserp_coordinatesretrievalstrategyname?: string;
  mserp_dataareaid_idname?: string;
  mserp_distancecalculationstrategyname?: string;
  mserp_fk_externallylocatedwarehousevendor_idname?: string;
  mserp_fk_mainrefillingwarehouse_idname?: string;
  mserp_fk_quarantinewarehouse_idname?: string;
  mserp_fk_transitwarehouse_idname?: string;
  mserp_formattedprimaryaddress?: string;
  mserp_identificationgroupname?: string;
  mserp_inventorystatuschangereservationremovallevelname?: string;
  mserp_inventprofiletype_runame?: string;
  mserp_isbillofladingprintingbeforeshipmentconfirmationenabledname?: string;
  mserp_isfallbackwarehousename?: string;
  mserp_isfinancialnegativeretailstoreinventoryallowedname?: string;
  mserp_ispalletmovementduringcyclecountingallowedname?: string;
  mserp_isphysicalnegativeretailstoreinventoryallowedname?: string;
  mserp_isprimaryaddressassignedname?: string;
  mserp_isrefilledfrommainwarehousename?: string;
  mserp_isretailstorewarehousename?: string;
  mserp_iswarehouseexternallymanagedname?: string;
  mserp_loadreleasereservationpolicyrulename?: string;
  mserp_primaryaddresstimezonename?: string;
  mserp_rawmaterialpickinginventoryissuestatusname?: string;
  mserp_reportasfinishedpostingmethodname?: string;
  mserp_shouldwarehouselocationidincludeaisleidname?: string;
  mserp_spatiallocationsortingstrategyname?: string;
  mserp_warehousereleasereservationrequirementrulefailureoptionname?: string;
  mserp_warehousereleasereservationrequirementrulename?: string;
  mserp_warehousetypename?: string;
  mserp_willautomaticloadreleasereserveinventoryname?: string;
  mserp_willinventorystatuschangeremoveblockingname?: string;
  mserp_willmanualloadreleasereserveinventoryname?: string;
  mserp_willorderreleasingconsolidateshipmentsname?: string;
  mserp_willproductionbomsreservewarehouselevelonlyname?: string;
  mserp_willshippingcancellationdecrementloadquanityname?: string;
  mserp_willwarehouselocationidincludebinidbydefaultname?: string;
  mserp_willwarehouselocationidincluderackidbydefaultname?: string;
  mserp_willwarehouselocationidincludeshelfidbydefaultname?: string;
  mserp_dataareaid_id?: object;
  _mserp_dataareaid_id_value?: string;
  mserp_fk_externallylocatedwarehousevendor_id?: object;
  _mserp_fk_externallylocatedwarehousevendor_id_value?: string;
  mserp_fk_mainrefillingwarehouse_id?: object;
  _mserp_fk_mainrefillingwarehouse_id_value?: string;
  mserp_fk_quarantinewarehouse_id?: object;
  _mserp_fk_quarantinewarehouse_id_value?: string;
  mserp_fk_transitwarehouse_id?: object;
  _mserp_fk_transitwarehouse_id_value?: string;
}